requirements.txt
python-dotenv==1.2.1
openai==2.16.0
langfuse==3.12.1
black==26.1.0
isort==7.0.0
requests==2.32.5
loguru==0.7.3
